Bounds : fixed values displayed in results as dirac (as well as in scene preview) Model class uses pointers to Events and Phases : easier to cast events to bounds Float maths functions used instead of double version Event Dialog title ok with bounds

+++ src/plugins/plugin_14C/Plugin14C.cpp
@@ -30,7 +30,7 @@ float Plugin14C::getLikelyhood(const float& t, const QJsonObject& data)
         float e = (curveG95Sup[t] - curveG[t]) / 1.96f;
         
         float variance = e * e + error * error;
-        float result = sqrt(variance) * exp(-0.5 * pow(g - age, 2) / variance);
+        float result = sqrtf(variance) * expf(-0.5 * powf(g - age, 2) / variance);
         
         return result;
     }
